pyproject.toml
[project]
name = "python-ai-agent"
version = "0.1.0"
description = "Русскоязычный ассистент по продажам, который отвечает на вопросы на основе предоставленного прайс-листа"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "langchain>=1.2.0",
    "langgraph-cli>=0.4.11",
    "langgraph-api>=0.6.13",
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "rapidfuzz>=3.14.3",
]

[dependency-groups]
dev = ["ipython>=8.0.0"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
